+++ packages/dx-react-chart/src/templates/series/area.test.tsx
@@ -2,11 +2,18 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 import { withStates } from '../../utils/with-states';
 import { withPattern } from '../../utils/with-pattern';
+import { withAnimation } from '../../utils/with-animation';
 import { Area } from './area';
+import {
+  getPathStart, processAreaAnimation, isCoordinatesChanged,
+} from '@devexpress/dx-chart-core';
 
 jest.mock('@devexpress/dx-chart-core', () => ({
   HOVERED: 'test_hovered',
   SELECTED: 'test_selected',
+  isCoordinatesChanged: jest.fn(),
+  getPathStart: jest.fn(),
+  processAreaAnimation: jest.fn(),
 }));
 
 jest.mock('../../utils/with-states', () => ({
@@ -14,6 +21,9 @@ jest.mock('../../utils/with-states', () => ({
 }));
 jest.mock('../../utils/with-pattern', () => ({
   withPattern: jest.fn().mockReturnValue(x => x),
+}));
+jest.mock('../../utils/with-animation', () => ({
+  withAnimation: jest.fn().mockReturnValue(x => x),
 }));
 
 describe('Area', () => {
@@ -25,7 +35,7 @@ describe('Area', () => {
     scales: { tag: 'test-scales' },
     rotated: true,
     clipPathId: 'clipPathId',
-    getAnimatedStyle: jest.fn(style => style),
+    style: { tag: 'test-style' },
   };
 
   it('should render root element', () => {
@@ -40,6 +50,7 @@ describe('Area', () => {
       fill: 'red',
       opacity: 0.5,
       clipPath: 'url(#clipPathId)',
+      style: { tag: 'test-style' },
     });
     expect(defaultProps.path).toBeCalledWith(defaultProps.coordinates);
   });
@@ -83,5 +94,28 @@ describe('Area', () => {
     ]);
     expect((withPattern as jest.Mock).mock.calls[0][0]({ index: 2 })).toEqual('series-2-hover');
     expect((withPattern as jest.Mock).mock.calls[1][0]({ index: 3 })).toEqual('series-3-selection');
+  });
+
+  it('should animate', () => {
+    expect(withAnimation)
+    .toBeCalledWith(processAreaAnimation, expect.any(Function), getPathStart, isCoordinatesChanged);
+  });
+
+  it('should update props', () => {
+    const tree = shallow((
+      <Area
+        {...defaultProps}
+      />
+    ));
+
+    tree.setProps({ ...defaultProps, coordinates: [4, 5, 6] });
+
+    expect(tree.find('path').props()).toEqual({
+      d: '4-5-6',
+      fill: 'red',
+      opacity: 0.5,
+      clipPath: 'url(#clipPathId)',
+      style: { tag: 'test-style' },
+    });
   });
 });