+++ packages/dx-react-chart/src/templates/series/path.tsx
@@ -1,18 +1,19 @@
 import * as React from 'react';
 import {
-  getAreaAnimationStyle, HOVERED, SELECTED,
+  processLineAnimation, HOVERED, SELECTED, isCoordinatesChanged, getPathStart,
 } from '@devexpress/dx-chart-core';
 import { withStates } from '../../utils/with-states';
+import { withAnimation } from '../../utils/with-animation';
 import { PathComponentPathProps } from '../../types';
 
 class RawPath extends React.PureComponent<PathComponentPathProps> {
   render() {
     const {
-      path,
+      path, animation,
       coordinates, rotated,
       index, state, pointComponent,
       color, clipPathId,
-      style, scales, getAnimatedStyle,
+      scales, pane,
       ...restProps
     } = this.props;
     return (
@@ -22,14 +23,18 @@ class RawPath extends React.PureComponent<PathComponentPathProps> {
         fill="none"
         strokeWidth={2}
         stroke={color}
-        style={getAnimatedStyle(style, getAreaAnimationStyle, scales)}
         {...restProps}
       />
     );
   }
 }
 
-export const Path = withStates({
+export const Path = withAnimation<any>(
+  processLineAnimation,
+  ({ coordinates }) => ({ coordinates }),
+  getPathStart,
+  isCoordinatesChanged,
+)(withStates({
   [HOVERED]: props => ({ strokeWidth: 4, ...props }),
   [SELECTED]: props => ({ strokeWidth: 4, ...props }),
-})(RawPath);
+})(RawPath));