+++ packages/dx-react-scheduler-material-ui/src/index.js
@@ -13,3 +13,5 @@ export * from './plugins/drag-drop-provider';
 export * from './plugins/today-button';
 export * from './plugins/edit-recurrence-menu';
 export * from './plugins/confirmation-dialog';
+export * from './plugins/current-time-indicator';
+export * from './plugins/resources';