+++ packages/dx-react-scheduler-material-ui/src/templates/all-day-panel/cell.jsx
@@ -9,7 +9,7 @@ import { getBorder } from '../utils';
 const styles = theme => ({
   cell: {
     padding: 0,
-    height: theme.spacing(7),
+    height: theme.spacing(5.75),
     borderLeft: getBorder(theme),
     '&:hover': {
       backgroundColor: theme.palette.action.hover,