+++ packages/dx-react-scheduler-material-ui/src/templates/drag-drop/appointments.jsx
@@ -3,35 +3,89 @@ import * as PropTypes from 'prop-types';
 import classNames from 'clsx';
 import Repeat from '@material-ui/icons/Repeat';
 import { POSITION_START, POSITION_END } from '@devexpress/dx-scheduler-core';
-import { withStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import { AppointmentContent } from '../appointment/appointment-content';
 import { Appointment } from '../appointment/appointment';
 import { SplitIndicator } from '../appointment/split-indicator';
-import { setColor } from '../utils';
+import { getAppointmentColor, getResourceColor } from '../utils';
 
-const draftStyles = theme => ({
+const draftStyles = makeStyles(theme => ({
   appointment: {
     boxShadow: theme.shadows[3],
     cursor: 'move',
     overflow: 'hidden',
-    backgroundColor: setColor(600, theme.palette.primary),
+    backgroundColor: resources => getAppointmentColor(
+      600, getResourceColor(resources), theme.palette.primary,
+    ),
     border: 0,
   },
-});
+  shadedAppointment: {
+    backgroundColor: resources => getAppointmentColor(
+      400, getResourceColor(resources), theme.palette.primary,
+    ),
+  },
+}));
 
-const sourceStyles = {
+const sourceStyles = makeStyles({
   appointment: {
     opacity: 0.5,
   },
+});
+
+export const DraftAppointment = ({
+  className, resources, isShaded, ...restProps
+}) => {
+  const classes = draftStyles(resources);
+  return (
+    <AppointmentBase
+      className={classNames({
+        [classes.appointment]: true,
+        [classes.shadedAppointment]: isShaded,
+      }, className)}
+      resources={resources}
+      {...restProps}
+    />
+  );
 };
 
-const DraftAppointmentBase = ({
-  classes, className, data, formatDate,
-  type, fromPrev, toNext, durationType, ...restProps
+DraftAppointment.propTypes = {
+  resources: PropTypes.array,
+  className: PropTypes.string,
+  isShaded: PropTypes.bool,
+};
+
+DraftAppointment.defaultProps = {
+  className: undefined,
+  resources: [],
+  isShaded: false,
+};
+
+export const SourceAppointment = ({ className, ...restProps }) => {
+  const classes = sourceStyles();
+  return (
+    <AppointmentBase
+      className={classNames(classes.appointment, className)}
+      {...restProps}
+    />
+  );
+};
+
+SourceAppointment.propTypes = {
+  className: PropTypes.string,
+};
+
+SourceAppointment.defaultProps = {
+  className: undefined,
+};
+
+const AppointmentBase = ({
+  className, data, formatDate, type, fromPrev,
+  toNext, durationType, isShaded, ...restProps
 }) => (
   <Appointment
-    className={classNames(classes.appointment, className)}
+    className={className}
     type={type}
+    isShaded={isShaded}
     {...restProps}
   >
     {fromPrev && <SplitIndicator position={POSITION_START} appointmentType={type} />}
@@ -46,8 +100,7 @@ const DraftAppointmentBase = ({
   </Appointment>
 );
 
-DraftAppointmentBase.propTypes = {
-  classes: PropTypes.object.isRequired,
+AppointmentBase.propTypes = {
   data: PropTypes.object.isRequired,
   fromPrev: PropTypes.bool.isRequired,
   toNext: PropTypes.bool.isRequired,
@@ -55,13 +108,12 @@ DraftAppointmentBase.propTypes = {
   durationType: PropTypes.string,
   className: PropTypes.string,
   type: PropTypes.string,
+  isShaded: PropTypes.bool,
 };
 
-DraftAppointmentBase.defaultProps = {
+AppointmentBase.defaultProps = {
   durationType: undefined,
   className: undefined,
   type: undefined,
+  isShaded: false,
 };
-
-export const DraftAppointment = withStyles(draftStyles, { name: 'DraftAppointment' })(DraftAppointmentBase);
-export const SourceAppointment = withStyles(sourceStyles, { name: 'SourceAppointment' })(DraftAppointmentBase);